fedreg-rag/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
langchain==0.0.267
langchain_community==0.0.10
langchain_openai==0.0.3
boto3==1.28.15
python-dotenv==1.0.0
chromadb==0.4.6
pydantic==1.10.8